AddPhotoButton, auth/verify 수정
